backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==1.10.13
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.0.0
